requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.9
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
attrs==24.2.0
beautifulsoup4==4.12.3
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
dnspython==2.7.0
fastapi==0.115.6
frozenlist==1.5.0
h11==0.14.0
httptools==0.6.4
idna==3.10
motor==3.6.0
multidict==6.1.0
neispy==5.0.2
propcache==0.2.1
pycomcigan==1.3.0
pydantic==2.10.3
pydantic_core==2.27.1
pymongo==4.9.2
python-dotenv==1.0.1
PyYAML==6.0.2
requests==2.32.3
sniffio==1.3.1
soupsieve==2.6
starlette==0.41.3
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.32.1
uvloop==0.21.0
watchfiles==1.0.0
websockets==14.1
yarl==1.18.3
